feat: 쿠키 삭제 로직 추가
