+++ src/main/java/com/hanghae/greenstep/member/MemberController.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.hanghae.greenstep.kakaoAPI.Dto.KakaoPhotoDto;
 import com.hanghae.greenstep.member.Dto.MemberRequestDto;
 import com.hanghae.greenstep.member.Dto.MemberResponseDto;
+import com.hanghae.greenstep.shared.CookieUtil;
 import com.hanghae.greenstep.shared.Message;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -12,12 +13,13 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/users")
 public class MemberController {
-
+    private final CookieUtil cookieUtil;
     private final MemberService memberService;
 
     @PatchMapping("/info")
@@ -36,6 +38,11 @@ public class MemberController {
     public ResponseEntity<?> getKakaoPhoto(HttpServletRequest request) throws JsonProcessingException {
         KakaoPhotoDto kakaoPhotoDto = memberService.getKakaoPhoto(request);
         return new ResponseEntity<>(Message.success(kakaoPhotoDto), HttpStatus.OK);
+    }
+
+    @GetMapping("/cookies")
+    public void deleteCookies(HttpServletRequest request, HttpServletResponse response){
+        cookieUtil.deleteAllCookies(request, response);
     }
 
 }
