+++ src/main/java/com/hanghae/greenstep/member/MemberService.java
@@ -26,7 +26,6 @@ import static com.hanghae.greenstep.shared.Status.DONE;
 import static com.hanghae.greenstep.shared.Status.WAITING;
 
 @Service
-@Slf4j
 @RequiredArgsConstructor
 public class MemberService {
     private final Check check;
@@ -52,19 +51,6 @@ public class MemberService {
         int missionCount = submitMissionRepository.countByMemberAndStatus(member, DONE);
         int waitingMissionCount = submitMissionRepository.countByMemberAndStatus(member, WAITING);
         return new MemberResponseDto(member, missionCount, waitingMissionCount);}
-
-    public void deleteAllCookies(HttpServletRequest request, HttpServletResponse response) {
-        Cookie[] cookies = request.getCookies(); // 모든 쿠키 가져오기
-        if (cookies != null) {
-            for (Cookie cookie : cookies) {
-                String name = cookie.getName(); // 쿠키 이름 가져오기
-                String value = cookie.getValue(); // 쿠키 값 가져오기
-                log.info("쿠키 이름 : " + name + "   쿠키값 : " + value);
-                cookie.setMaxAge(0); // 유효시간을 0으로 설정
-                response.addCookie(cookie);
-            }
-        }
-    }
 
     public KakaoPhotoDto getKakaoPhoto(HttpServletRequest request) throws JsonProcessingException {
         String accessToken = request.getHeader("Kakao_Authorization");
